This PR introduces separate column collector helpers for PgHeapTuple and Vector<PgHeapTuple> types to improve performance dramatically. The performance penalty is while running
```
for tuple in tuples {
  let att: Option<PgHeapTuple> = tuple.get_by_name(att_name);
  ...
}
```

That was because pgrx calls 'lookup_rowtype_tupdesc' every time when converting the attribute 'Datum'
into PgHeapTuple. This causes significant performance penalty when we iterate over millions of tuples
to extract the same column. To prevent that from occuring, we needed to share a single TupleDesc for
the attribute of the tuples and did not have to create a new one every time.